+++ tiberiandawn/cell.h
@@ -261,7 +261,7 @@ public:
     /*
     **	Maintenance calculation support.
     */
-    long Tiberium_Adjust(bool pregame = false);
+    int Tiberium_Adjust(bool pregame = false);
     void Wall_Update(void);
     void Concrete_Calc(void);
     void Recalc_Attributes(void);
