+++ tiberiandawn/globals.cpp
@@ -189,7 +189,7 @@ void* SpeechBuffer;
 ** This accumulates into a useful value that contributes to a
 **	histogram of game performance.
 */
-long SpareTicks;
+int SpareTicks;
 
 /***************************************************************************
 **	This is a special scenario count down value. End of game condition will
@@ -274,7 +274,7 @@ CreditClass CreditDisplay;
 **	has been displayed. Once the message has been displayed, it will not be
 **	displayed again.
 */
-long TutorFlags[2];
+int TutorFlags[2];
 
 /**************************************************************************
 ** This class records the special command override options that C&C
@@ -347,7 +347,7 @@ bool GameActive;
 **	a long, but the value wasn't supplied to a function. This is used
 **	specifically for the default reference value. As such, it is not stable.
 */
-long LParam;
+int LParam;
 
 #ifdef SCENARIO_EDITOR
 /***************************************************************************
@@ -739,7 +739,7 @@ const char* SerialPacketNames[] = {
 /***************************************************************************
 **	These variables are just to help find sync bugs.
 */
-long TrapFrame = 0x7fffffff;        // frame to start trapping object values at
+int TrapFrame = 0x7fffffff;         // frame to start trapping object values at
 RTTIType TrapObjType = RTTI_NONE;   // type of object to trap
 TrapObjectType TrapObject = {NULL}; // ptr to object being trapped
 COORDINATE TrapCoord = 0;           // COORD of object to trap
@@ -829,7 +829,7 @@ int MetaSize = ((546 - sizeof(CommHeaderType)) / sizeof(EventClass)) * sizeof(Ev
 ** multiplayer games.
 */
 int Seed = 0;
-long* RandSeedPtr;
+int* RandSeedPtr;
 
 /***************************************************************************
 ** If this value is non-zero, use it as the random # seed instead; this should
